[Silver I] Title: 쉬운 계단 수, Time: 208 ms, Memory: 17716 KB -BaekjoonHub

+++ 백준/Silver/10844. 쉬운 계단 수/쉬운 계단 수.java
@@ -6,7 +6,7 @@ public class Main {
 	public static void main(String[] args) {
 		Scanner sc = new Scanner(System.in);
 		int N = sc.nextInt();
-		long[][] dp = new long[10][N + 1];
+		int[][] dp = new int[10][N + 1];
 		long count = 0;
 		
 		for (int num = 1; num <= 9; num++) {
